remove rename new mcts.py
